Added command history feature
This commit adds a command history feature.  It works similarly to other command shells, where the up and down arrow keys cycle through recently used commands. The user can make changes to a previous command's text before running it again.  If the user has already entered part of a command and then scrolls through the history, the partial command is saved at the beginning of the history list.we  This feature requires a significant amount of memory to be used for storing the command history list.  As a rule of thumb, command history should only be enabled on systems with at least 8k of RAM (Arduino MEGA and better).  It can be enabled on smaller systems, but only when (CONFIG_SHELL_MAX_INPUT * (CONFIG_SHELL_COMMAND_HISTORY + 1)) + 4 bytes of RAM are available.  For a history list that is 10 commands long and the default input buffer length 0f 70, this would be 774 bytes.  This much memory usage only makes sense on some systems, so this feature is disabled by default.  To enable it, set the CONFIG_SHELL_COMMAND_HISTORY macro in Shell.h to the desired command history list length.  The max history length is 255 entries.

To implement this feature, the following changes were made:
--A number of global variables were added (4 bytes), and count was moved from a static in shell_task to be a global as well
--shellbuf is expanded into a two dimensional array, the first dimension is the list of historic commands, and the second dimension is the index within that command buffer.  The command history list is implemented as a ring buffer.
--Added shell_clear_command(), a public function to clear the current command buffer and displayed text in the shell
--Basic processing of VT100 escape sequences, including Command Sequence Indroducer sequences, for receiving arrow keys.  This may also lead to implementing the left and right arrow keys as a future project, allowing the user to move a cursor left and right when typing a command.  The biggest hurdle to this right now is that editing the middle or beginning of a long command will lead to lots of inserting in the middle of a command buffer, which will be slow.
--In addition to using count to track the length of an entered command, command text (including arguments) are now terminated by a NULL char
--When command history is enabled, command text (including arguments) are now copied to a new buffer for parsing.  This prevents the NULL string terminators between each argument from being added to the command history entry.