Added shell_set_prompt()
Added a funtion to set the command prompt text, and a call to this function in s
hell_init() to initialize the prompt text to "device".  NOTE: This function expe
cts to receive a pointer to a string in RAM, not PROGMEM.  If a PROGMEM string i
s desired, first copy it to RAM with strcpy_P().