Added shell_refresh()
This PR adds shell_refresh(), which refreshes the command prompt by removing any input from the input buffer and shell, and prints a welcome message and new command prompt.  This is useful if the user will be connecting to the shell after the command prompt has been printed (eg, when connecting via Telnet).