Make output buffer size easily set by user
This PR creates a macro CONFIG_SHELL_OUT_BUFFER_LEN which defines the
length of the output buffer.  This should be large enough to accommodate
most large writes via the shell_print functions, but still small enough
to fit in a single shell_bwriter call (ie, for Telnet, smaller than the
TCP Maximum Transmission Unit, 1500 bytes).  This buffer is also
statically allocated, and should not be so large that it uses all the
available RAM in the system.  The macro is set to the original output
buffer size (30 bytes).